Import working version of passport.py
Shows live camera view. Automatically centers and zooms image to be
precisely correct for passport photos. Hit spacebar to snap the photo
to "passport.jpg" and 'q' to quit.

Currently, size is hardcoded to the Chinese standard, 33mm x 48mm,
with the distance from the chin to the bottom of the photo being 7mm
and the eyes centered.