Don't try to set camera to maximum resolution
Using an arbitrarily high resolution used to work. Today it sets the
resolution to 0x0, which of course fails.

It looks like OpenCV can only enumerate the video modes for certain
camera backends, but not all of them. So, the best practice is to just
leave the resolution alone.

L'sigh.