More explicit information about possible bugs
Test of High-level GUI bugs in OpenCV 4.5 and 4.6.
Keys:
        change the image size:          1 to 9, or 0 for vidcap size.
        resize window to image size:    r
        toggle fullscreen:              f
        get window properties           g
        resize then enable fullscreen   t

Interaction between fullscreen and resize is unclear and likely buggy.

In 4.6, resizing the window while in fullscreen will turn off
fullscreen. It ought to stay in fullscreen and remember the new window
size for when fullscreen is turned off. Additionally, attempting to
programmatically work around OpenCV turning off fullscreen by turning
it back on fails, likely because OpenCV's resizeWindow is running
asynchronously..

To test:

1. Press 5 r and notice that resizing the window to 500x500
   works correctly when not fullscreen.

2. Press f 6 r and notice that as soon as resizeWindow() is called,
   fullscreen was turned off. It should have stayed fullscreen.

3. Now try f b, which attempts to both resize the window and enable
   fullscreen mode immediately afterward. Note that it fails: the
   screen stays windowed, probably due to a race with some
   asynchronous GUI process caused by resizeWindow().

In 4.5, windows cannot be made resized smaller. To test, press 5 r 6 r
and notice that the window can resize larger. Now press 4 r 3 r, the
window fails to get smaller.