Add option to change codec and fps of camera
While it turned out not to be helpful, I wanted to see if setting the
codec to MJPG on an old USB 2 camera would give a better picture since
it would have twice the resolution. It turned out, no because of the
quality degradation from the compression. I tried lowering the FPS in
the hope that it would use less severe compression, but that didn't
help.

Nevertheless, I'll leave these options in the program in case someone
else wants to try them.