fix weird zooming caused by left eye == right eye bug
Would on rare occasions have a bug where suddenly the image would zoom
in the eye. The left eye and right eye were getting detected as being
at essentially the same location, which of course is wrong. I had put
in a filter filter that ensured they were on opposite sides of the
eye_pair, but it was possible for that to get fooled. For example:

    left=(107,84)
    right=(120,82)
    eye_pair=(44,82,145,36)  # (x,y,w,h)

As a fix, I've changed isWithinLeft() / Right() to only accept an eye
that is within the first / last _quarter_ of the eye_pair rectangle
instead of half.

Todo: maybe also check if the center of either eye is within the
box returned by the haar cascade for the other.