Improve eye accuracy by using height from eye_pair
The Haar cascades for individual eyes, haarcascade_{left,right}eye_2splits.xml,
are not ideal for this program. They include the eyebrow in their bounding box.

A simple fix is to set the height of each eye to the height of
the eye_pair cascade, haarcascade_mcs_eyepair_big.xml, which is much
more accurate.

Of course, this is presuming the eyes are approximately horizontal,
but that's already necessary for the Haar cascades anyhow.