Fix iod bug: should use maxpect for desired distances
Was previously using the image width and height, but that's just a
scaled down version of the camera. Instead we should use the
photo_aspect ratio to calculate the final width and final height that
the image will be displayed at and use those for eye_distance and
eye_height.